+ Tutorial 4 : Added CMakeLists - Corrected mouse handling in computeMatricesFromInputs : does not depend on fps * Clearer objloader * VBO indexer : moved from unsigned int to unsigned short back again + Tutorial 15 : initial commit

+++ common/vboindexer.hpp
@@ -6,7 +6,7 @@ void indexVBO(
 	std::vector<glm::vec2> & in_uvs,
 	std::vector<glm::vec3> & in_normals,
 
-	std::vector<unsigned int> & out_indices,
+	std::vector<unsigned short> & out_indices,
 	std::vector<glm::vec3> & out_vertices,
 	std::vector<glm::vec2> & out_uvs,
 	std::vector<glm::vec3> & out_normals
